Move trait `IntoQueryParam` into its own file
Removing coherent and self-contained chunks of code from a catch-all "plumbing"
module is typically an improvement.

Giving the trait its own file also gets rid of the unhelpful `sealed` module
that it was previously in.
